requirements.txt
Django==1.9.8
numpy
Django==1.9.8
numpy==1.13.1
olefile==0.44
opencv-python==3.3.0.10
Pillow==4.2.1
scipy==0.19.1
